backend/requirements.txt
pandas
selenium
yfinance
psycopg2-binary
SQLAlchemy
dash
plotly
python-dotenv
pandas
statsmodels
scikit-learn
matplotlib
pmdarima
statsmodels
dash-extensions
flask
flask-cors
webdriver-manager